fxbot/collectors/requirements.txt
sqlalchemy==2.0.36
asyncpg==0.29.0
python-dotenv==1.0.1
httpx==0.28.1
apscheduler==3.10.4